reestructure structs
