+++ src/api/models/mod.rs
@@ -1,0 +1,3 @@
+pub mod compute;
+pub mod inference;
+pub mod pipeline;
