+++ src/lib.rs
@@ -1,6 +1,3 @@
 pub mod api;
-pub mod compute;
 pub mod cu29;
-pub mod inference;
 pub mod pipeline;
-pub mod stats;
